Refactors download logic from retrieval page to download table
