+++ components/form-builder/app/DownloadTable/DownloadTable.tsx
@@ -1,4 +1,4 @@
-import React from "react";
+import React, { useEffect, useState, useRef } from "react";
 import { VaultSubmissionList } from "@lib/types";
 import { useTranslation } from "react-i18next";
 import { SkipLinkReusable } from "@components/globals/SkipLinkReusable";
@@ -6,6 +6,11 @@ import { ConfirmReceiptStatus } from "./ConfirmReceiptStatus";
 import { DownloadResponseStatus } from "./DownloadResponseStatus";
 import { RemovalStatus } from "./RemovalStatus";
 import { DownloadStatus } from "./DownloadStatus";
+import { useRouter } from "next/router";
+import { logMessage } from "@lib/logger";
+import axios from "axios";
+import { ToastContainer, toast } from "react-toastify";
+import "react-toastify/dist/ReactToastify.min.css";
 
 export enum VaultStatus {
   NEW = "New",
@@ -15,11 +20,8 @@ export enum VaultStatus {
 }
 
 interface DownloadTableProps {
-  submissions: VaultSubmissionList[];
-  selectionStatus: Map<string, boolean>;
-  setSelectionStatus: (newStatus: Map<string, boolean>) => void;
-  checkedItems: Map<string, boolean>;
-  handleDownload: React.MouseEventHandler<HTMLButtonElement>;
+  vaultSubmissions: VaultSubmissionList[];
+  formId?: string;
 }
 
 export function getDaysPassed(date: Date): number {
@@ -30,20 +32,97 @@ export function getDaysPassed(date: Date): number {
   return daysPassed;
 }
 
-export const DownloadTable = ({
-  submissions,
-  selectionStatus,
-  setSelectionStatus,
-  checkedItems,
-  handleDownload,
-}: DownloadTableProps) => {
+export const DownloadTable = ({ vaultSubmissions, formId }: DownloadTableProps) => {
   const { t } = useTranslation("form-builder");
+  const router = useRouter();
+  const MAX_FILE_DOWNLOADS = 20;
+  const toastPosition = toast.POSITION.TOP_CENTER;
+  const checkedItems = useRef(new Map());
+  const [selectionStatus, setSelectionStatus] = useState(
+    new Map(vaultSubmissions.map((submission) => [submission.name, false]))
+  );
+
+  useEffect(() => {
+    const checkedMap = new Map();
+    selectionStatus.forEach((checked, name) => {
+      if (checked) {
+        checkedMap.set(name, checked);
+      }
+    });
+    checkedItems.current = checkedMap;
+  }, [selectionStatus]);
 
   const handleChecked = (e: React.ChangeEvent<HTMLInputElement>) => {
     const name = e.target.id;
     const checked: boolean = e.target.checked;
     // Note: "new Map" Needed to clone and set so React change detection notices a change in the Map
     setSelectionStatus(new Map(selectionStatus.set(name, checked)));
+  };
+
+  // NOTE: browsers have different limits for simultaneous downloads. May need to look into
+  // batching file downloads (e.g. 4 at a time) if edge cases/* come up.
+  const handleDownload = async () => {
+    if (checkedItems.current.size === 0) {
+      toast.warn(t("downloadResponsesTable.download.atLeastOneFile"), { position: toastPosition });
+      return;
+    }
+
+    if (checkedItems.current.size > MAX_FILE_DOWNLOADS) {
+      toast.warn(
+        t("downloadResponsesTable.download.trySelectingLessFiles", { max: MAX_FILE_DOWNLOADS }),
+        { position: toastPosition }
+      );
+      return;
+    }
+
+    const toastDownloadingId = toast.info(
+      t("downloadResponsesTable.download.downloadingXFiles", {
+        fileCount: checkedItems.current.size,
+      }),
+      { position: toastPosition, autoClose: false }
+    );
+
+    try {
+      const downloads = Array.from(checkedItems.current, async ([submissionName]) => {
+        if (!submissionName) {
+          throw new Error("Error downloading file from Retrieval table. SubmissionId missing.");
+        }
+        const url = `/api/id/${formId}/${submissionName}/download`;
+        const fileName = `${submissionName}.html`;
+        return axios({
+          url,
+          method: "GET",
+          responseType: "blob",
+          timeout: process.env.NODE_ENV === "production" ? 60000 : 0,
+        }).then((response) => {
+          const url = window.URL.createObjectURL(new Blob([response.data]));
+          const link = document.createElement("a");
+          link.href = url;
+          link.setAttribute("download", fileName);
+          document.body.appendChild(link);
+          link.click();
+        });
+      });
+
+      await Promise.all(downloads).then(() => {
+        // TODO: Future tech debt. See https://github.com/cds-snc/platform-forms-client/issues/1744
+        setTimeout(() => {
+          // Refreshes getServerSideProps data without a full page reload
+          router.replace(router.asPath);
+          toast.dismiss(toastDownloadingId);
+          toast.success(t("downloadResponsesTable.download.downloadComplete"), {
+            position: toastPosition,
+          });
+        }, 400);
+      });
+    } catch (err) {
+      logMessage.error(err as Error);
+      toast.dismiss(toastDownloadingId);
+      toast.error(t("downloadResponsesTable.download.errorDownloadingFiles"), {
+        position: toastPosition,
+        autoClose: false,
+      });
+    }
   };
 
   return (
@@ -72,7 +151,7 @@ export const DownloadTable = ({
           </tr>
         </thead>
         <tbody>
-          {submissions.map((submission) => (
+          {vaultSubmissions.map((submission) => (
             <tr
               key={submission.name}
               className={
@@ -138,9 +217,14 @@ export const DownloadTable = ({
           aria-live="polite"
         >
           {t("downloadResponsesTable.download.downloadXSelectedResponses", {
-            size: checkedItems.size,
+            size: checkedItems.current.size,
           })}
         </button>
+      </div>
+
+      {/* Sticky position to stop the page from scrolling to the top when showing a Toast */}
+      <div className="sticky top-0">
+        <ToastContainer />
       </div>
     </section>
   );
